third_party/blink/web_tests/external/wpt/css/css-values/calc-serialization-002-expected.txt

This is a testharness.js-based test.
[FAIL] testing calc(1vh + 2px + 3%)
  assert_equals: expected "calc(3% + 2px + 1vh)" but got "calc(1vh + 2px + 3%)"
[PASS] testing calc(4px + 1vh)
[FAIL] testing calc(5px + 6em + 1vh)
  assert_equals: expected "calc(6em + 5px + 1vh)" but got "calc(5px + 6em + 1vh)"
[FAIL] testing calc(-8px + 9em + 1vh)
  assert_equals: expected "calc(9em - 8px + 1vh)" but got "calc(-8px + 9em + 1vh)"
[FAIL] testing calc(1pc + 1in + 1vh + 10%)
  assert_equals: expected "calc(10% + 112px + 1vh)" but got "calc(112px + 1vh + 10%)"
[FAIL] testing calc(25.4q + 1vh + 12%)
  assert_equals: expected "calc(12% + 24px + 1vh)" but got "calc(24px + 1vh + 12%)"
[FAIL] testing calc(1em + 1.27cm + 13% + 3em)
  assert_equals: expected "calc(13% + 4em + 48px)" but got "calc(4em + 48px + 13%)"
[FAIL] testing calc(15vw + 16vmin - 17vh)
  assert_equals: expected "calc(-17vh + 16vmin + 15vw)" but got "calc(15vw + 16vmin - 17vh)"
[PASS] testing calc(9pt + calc(9rem + 10px))
[FAIL] testing calc(5pt + 5em + 4pt + 3em)
  assert_equals: expected "calc(8em + 12px)" but got "calc(12px + 8em)"
[FAIL] testing calc(4vmin + 0pt + 3pc)
  assert_equals: expected "calc(48px + 4vmin)" but got "calc(4vmin + 48px)"
[FAIL] testing calc(4vmin + 0pt)
  assert_equals: expected "calc(0px + 4vmin)" but got "calc(4vmin + 0px)"
[PASS] testing calc(100% - 100% + 1em)
[PASS] testing calc(100% + 1em - 100%)
[FAIL] testing calc(1vh - 7px)
  assert_equals: expected "calc(-7px + 1vh)" but got "calc(1vh - 7px)"
[PASS] testing calc(5ex - 9ex)
[PASS] testing calc(-80px + 25.4mm)
[PASS] testing calc(2 * (10px + 1rem))
[PASS] testing calc(2 * (10px - 1rem))
[PASS] testing calc((10px + 1rem) / 2)
[PASS] testing calc(2 * (min(10px, 20%) + max(1rem, 2%)))
[PASS] testing calc((min(10px, 20%) + max(1rem, 2%)) * 2)
[FAIL] testing calc(1vmin - 14%)
  assert_equals: expected "calc(-14% + 1vmin)" but got "calc(1vmin - 14%)"
[PASS] testing calc(4 * 3px + 4pc / 8)
Harness: the test ran to completion.
